fix(cosh-ng): [shell] restore layout gates
- Extract raw-input models and relay helpers into owner submodules.
- Refresh inventory baselines after concurrent main changes.
- Preserve runtime behavior while restoring CI gate coverage.
