+++ src/cosh-ng/crates/cosh-shell/src/shell_host/raw_relay/terminal_size.rs
@@ -1,0 +1,35 @@
+//! Outer-terminal size synchronization for the shell PTY.
+
+use std::io;
+
+use nix::libc;
+use nix::pty::Winsize;
+
+use crate::raw_input::{set_pty_winsize, signal_process_group};
+
+use super::super::model::current_terminal_winsize;
+
+pub(super) fn sync_outer_terminal_winsize(
+    master_fd: i32,
+    child_pid: u32,
+    last_winsize: &mut Winsize,
+) -> io::Result<()> {
+    let Some(current) = current_terminal_winsize() else {
+        return Ok(());
+    };
+    if same_winsize(&current, last_winsize) {
+        return Ok(());
+    }
+
+    set_pty_winsize(master_fd, current)?;
+    signal_process_group(child_pid, libc::SIGWINCH)?;
+    *last_winsize = current;
+    Ok(())
+}
+
+fn same_winsize(left: &Winsize, right: &Winsize) -> bool {
+    left.ws_row == right.ws_row
+        && left.ws_col == right.ws_col
+        && left.ws_xpixel == right.ws_xpixel
+        && left.ws_ypixel == right.ws_ypixel
+}
