+++ src/cosh-ng/scripts/check-test-inventory.sh
@@ -53,8 +53,8 @@ check_overlap_ceiling() {
 check_source_floor cosh-types 24
 check_source_floor cosh-platform 279
 check_source_floor cosh-cli 71
-check_source_floor cosh-core 626
-check_source_floor cosh-shell 2371
+check_source_floor cosh-core 658
+check_source_floor cosh-shell 2467
 
 ignored_count="$(rg -n '^[[:space:]]*#\[ignore' crates -g '*.rs' | wc -l | tr -d ' ')"
 echo "ignored tests: $ignored_count"
@@ -63,7 +63,7 @@ if [[ "$ignored_count" -ne 3 ]]; then
 fi
 
 check_overlap_ceiling cosh-core cosh-core 4
-check_overlap_ceiling cosh-shell cosh-shell 613
+check_overlap_ceiling cosh-shell cosh-shell 620
 
 if [[ "$failures" -ne 0 ]]; then
   echo "test inventory audit failed with $failures violation(s)" >&2
